memasukkan statement if ke dalam REST PUT

+++ src/main/java/atmwsc/wscrestful/ProductServiceController.java
@@ -51,11 +51,15 @@ public class ProductServiceController {
 
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
-   }   
+       if(!productRepo.containsKey(id)){
+            return new ResponseEntity<>("Id Product belum ada, silakan edit Id yang sudah ada!", HttpStatus.NOT_FOUND);
+       }
+       
+       productRepo.remove(id);
+            product.setId(id);
+            productRepo.put(id, product);
+            return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+    }   
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
    public ResponseEntity<Object> delete(@PathVariable("id") String id) {
